remove unnecessary auth
