+++ api/src/handlers/auth/routes.rs
@@ -1,0 +1,7 @@
+use actix_web::{Scope, web};
+use crate::handlers::auth::handler;
+
+pub fn routes() -> Scope {
+    web::scope("/auth")
+        .service(handler::login)
+}
