+++ api/src/handlers/mod.rs
@@ -1,5 +1,5 @@
 mod drinks;
 mod tags;
 pub mod routes;
-mod users;
+mod auth;
 mod types;
