--- api/src/handlers/users/handler.rs
@@ -1,30 +1,0 @@
-use actix_web::{get, HttpResponse, post, Responder, web};
-use argon2::{Argon2, PasswordHash, PasswordVerifier};
-use crate::extractors::claims::Claims;
-use crate::handlers::types::Dto;
-use crate::handlers::users::service;
-use crate::handlers::users::types::{CreateUserRequest, LoginUserRequest, UserDto};
-
-#[get("/me")]
-pub async fn get(claims: Claims) -> impl Responder {
-    HttpResponse::Ok().json(claims)
-}
-
-#[post("")]
-pub async fn create(request: web::Json<CreateUserRequest>) -> impl Responder {
-    let data = service::create(request.0).await.unwrap();
-    let user = UserDto::from_model(data);
-    HttpResponse::Ok().json(user)
-}
-
-#[post("/login")]
-pub async fn login(request: web::Json<LoginUserRequest>) -> impl Responder {
-    if let Some(user) = service::get_from_email(request.0.email).await {
-        let parsed = PasswordHash::new(&user.password).unwrap();
-        if Argon2::default().verify_password(request.0.password.as_bytes(), &parsed).is_ok() {
-            let response = service::get_token().await;
-            return HttpResponse::Ok().json(response);
-        }
-    }
-    HttpResponse::BadRequest().body("credentials are invalid")
-}
