--- api/src/handlers/users/routes.rs
@@ -1,9 +1,0 @@
-use actix_web::{Scope, web};
-use crate::handlers::users::handler;
-
-pub fn routes() -> Scope {
-    web::scope("/users")
-        .service(handler::get)
-        .service(handler::create)
-        .service(handler::login)
-}
